requirements.txt
aiofiles==23.2.1
alembic==1.11.2
annotated-types==0.5.0
anyio==4.4.0
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.2
Cython==3.0.10
databases==0.9.0
dnspython==2.6.1
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.100.0
FastAPI-SQLAlchemy==0.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
numpy==2.0.0
orjson==3.10.5
pandas==2.2.2
passlib==1.7.4
pillow==10.3.0
psycopg2==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.0.3
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
reportlab==4.2.2
requests==2.31.0
rsa==4.9
setuptools==70.1.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
stripe==8.10.0
typing_extensions==4.7.1
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.23.0
watchfiles==0.19.0
websockets==11.0.3
wheel==0.43.0
